TestBackwardsCompatibility code has been edited to include the code that was used to create the old index
To run the testCreateNewTaxonomy function:
ant test -Dtestcase=TestBackwardsCompatibility
-Dtests.bwcdir=/Users/gauworah/oldTaxoIndexOct  -Dtests.codec=default
-Dtests.useSecurityManager=false -Dtestmethod=testCreateNewTaxonomy

When I run the testCreateNewTaxonomy for the first time it fails with an
error on L98 indicating that it is unable to read the new `b`
BinaryDocValues field. This is the correct expected behavior.

However, when I run it again on the same directory which contains the
old index, it succeeds i.e., the build passes.